src/requirements.txt
biopython @ file:///home/conda/feedstock_root/build_artifacts/biopython_1704895409820/work
easydict==1.9
fair-esm==2.0.0
huggingface-hub==0.22.2
networkx==3.1
ninja==1.11.1.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808301083/work
pillow @ file:///croot/pillow_1707233021655/work
pyg-lib==0.4.0+pt22cu121
rdkit==2024.3.1
scikit-learn==1.3.2
torch==2.2.2
torch_cluster==1.6.3+pt22cu121
torch_geometric==2.4.0
torch_scatter==2.1.2+pt22cu121
torch_sparse==0.6.18+pt22cu121
torch_spline_conv==1.2.2+pt22cu121
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
unicore @ git+https://github.com/dptech-corp/Uni-Core.git@3ee5bbadbf1eaa1e52222bb8e864c0949aa7ca94
vina==1.2.5
